rename source files for efixplorer
