+++ efiXplorer/efi_analyser_x86.cc
@@ -20,13 +20,13 @@
  *
  */
 
-#include "efiAnalyzer.h"
-#include "efiGlobal.h"
-#include "efiUi.h"
-#include "efiUtils.h"
+#include "efi_analyser.h"
+#include "efi_global.h"
+#include "efi_ui.h"
+#include "efi_utils.h"
 
 #ifdef HEX_RAYS
-#include "efiHexRays.h"
+#include "efi_hexrays.h"
 #endif
 
 using namespace efi_analysis;
